Move group to models
